Formatting and rearranging packages
