+++ ui/src/main/java/com/neaniesoft/vermilion/ui/images/ImageRouter.kt
@@ -1,12 +1,9 @@
 package com.neaniesoft.vermilion.ui.images
 
 import android.net.Uri
+import com.neaniesoft.vermilion.ui.images.matchers.ImageHostMatchResult
+import com.neaniesoft.vermilion.ui.images.matchers.ImageHostMatcher
 import com.neaniesoft.vermilion.utils.logger
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import dagger.multibindings.IntoSet
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -24,33 +21,4 @@ class ImageRouter @Inject constructor(private val matchers: Set<@JvmSuppressWild
         }
         return null
     }
-}
-
-interface ImageHostMatcher {
-    fun match(uri: Uri): ImageHostMatchResult
-}
-
-sealed class ImageHostMatchResult {
-    object NoMatch : ImageHostMatchResult()
-    data class DirectImageUri(val uri: Uri) : ImageHostMatchResult()
-}
-
-@Singleton
-class RedditImageHostMatcher @Inject constructor() : ImageHostMatcher {
-    override fun match(uri: Uri): ImageHostMatchResult {
-        return if (uri.host == "i.redd.it") {
-            ImageHostMatchResult.DirectImageUri(uri)
-        } else {
-            ImageHostMatchResult.NoMatch
-        }
-    }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class ImageHostMatcherModule {
-
-    @Binds
-    @IntoSet
-    abstract fun provideImageHostMatchers(redditImageHostMatcher: RedditImageHostMatcher): ImageHostMatcher
 }
