+++ ui/src/main/java/com/neaniesoft/vermilion/ui/images/matchers/ImageHostMatchResult.kt
@@ -1,0 +1,8 @@
+package com.neaniesoft.vermilion.ui.images.matchers
+
+import android.net.Uri
+
+sealed class ImageHostMatchResult {
+    object NoMatch : ImageHostMatchResult()
+    data class DirectImageUri(val uri: Uri) : ImageHostMatchResult()
+}
