+++ ui/src/main/java/com/neaniesoft/vermilion/ui/images/matchers/ImageHostMatcher.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.ui.images.matchers
+
+import android.net.Uri
+
+interface ImageHostMatcher {
+    fun match(uri: Uri): ImageHostMatchResult
+}
